fix: use parse float instead of parse int for package price

+++ src/app/(components)/PackageBoxCard.tsx
@@ -48,7 +48,9 @@ export default function PackageBoxCard({ item }: { item: PackagesDataType }) {
             <span className="text-custom-blue font-medium">
               {item.currency === 'USD' ? '$' : 'NPR.'}
               {item.price
-                ? parseInt(item.price).toFixed(item.currency === 'USD' ? 2 : 0)
+                ? parseFloat(item.price).toFixed(
+                    item.currency === 'USD' ? 2 : 0
+                  )
                 : '-'}
             </span>
             /p
